relationships

+++ routes/web.php
@@ -21,28 +21,34 @@ use App\Http\Controllers\PostsController;
 Route::get('/', [PostsController::class,'index']);
 
 // show create form
-Route::get('/posts/create',[PostsController::class,'create']);
+Route::get('/posts/create',[PostsController::class,'create'])->middleware('auth');
 
 // store data
-Route::post('/posts',[PostsController::class,'store']);
+Route::post('/posts',[PostsController::class,'store'])->middleware('auth');
 
 // show edit form
-Route::get('/posts/{Post}/edit',[PostsController::class,'edit']);
+Route::get('/posts/{Post}/edit',[PostsController::class,'edit'])->middleware('auth');
 
 // update form
-Route::put('/posts/{Post}',[PostsController::class,'update']);
+Route::put('/posts/{Post}',[PostsController::class,'update'])->middleware('auth');
 
 // update form
-Route::delete('/posts/{Post}',[PostsController::class,'destroy']);
+Route::delete('/posts/{Post}',[PostsController::class,'destroy'])->middleware('auth');
 
 // show single post
 Route::get('/posts/{id}',[PostsController::class,'show']);
 
-//register 
-Route::get('/register',[UserController::class,'create']);
+// show register  form
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
 
 //create new user
-Route::post('/users',[UserController::class,'store']);
+Route::post('/users',[UserController::class,'store'])->middleware('guest');
 
 // logout user 
-Route::post('/logout',[UserController::class,'logout']);
+Route::post('/logout',[UserController::class,'logout'])->middleware('auth');
+
+// show user login form
+Route::get('/login',[UserController::class,'login'])->name('login');
+
+// login user
+Route::post('/users/authenticate',[UserController::class,'authenticate']);
